Added New Updates (Create Poll, Update User, Curr User)
ADDED NOTES
Added the Auth to keep track of current logged in user
Added the Auth.js to flow with all files so the current user can be known on all pages.
Added the Create Poll System
Added a way to display logged in user and statistics on Home Page
Added a way to turn off login and register modals when not needed.
Added Protection to avoid starting polls before current date, also not to allow the endDate to be less than the startTime
Added Protection if the current day is the same as Start Date then compare time so see if Active or Inactive and to avoid incorrect Poll Formats.
Added update to Account Doc to update the values of Polls left, Inactive and Active polls based on creation.
Added Photos for Reference

NEEDED NOTES
Need to add a way to turn polls to active once the start date is above current date.
Need to add the functionality of changing a Poll element to the correct values based on the Poll
Need to add the way of getting a one time passcode
Need to fix and modify more protection to correct Poll Time Frame Formats
Need to fix documentation and modify code to make more ease, and improve readability.