Update: MyPolls, FilterOnPolls, Fixed Selection, and Highlights
Added
PollContainer Class - Holds the polls related to the webpage that is being viewed. currently uses the old strategy of showing the polls but increases writes, this is a rough draft and there is a better idea which will be implemented.
UserDash.js - Now shows the owners their polls (Inactive or Active)
Polls.html - Shows two buttons for seeing Active Polls and Voted On Polls with Polls.js implementing the functionality.
Vote.js - All features related to voting have been moved here to reduce clutter.

Fixed
* No longer adds multiple votes to the same selectionName